pyproject.toml
[project]
name = "ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "chromadb>=1.5.5",
    "cloudinary>=1.44.1",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.6",
    "motor>=3.7.1",
    "openai>=2.26.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
